Move base wdio.conf to tests directory
