+++ tests/e2e/default-mobile/wdio.conf.js
@@ -1,4 +1,4 @@
-const wdioBaseConfig = require('../wdio.conf');
+const wdioBaseConfig = require('../../wdio.conf');
 
 const chai = require('chai');
 chai.use(require('chai-exclude'));
