+++ tests/e2e/default/wdio.conf.js
@@ -1,4 +1,4 @@
-const wdioBaseConfig = require('../wdio.conf');
+const wdioBaseConfig = require('../../wdio.conf');
 const { suites } = require('./suites');
 
 // Override specific properties from wdio base config
